refactor a few public members out of DataSchemaRichContextTraverser
